Agent/requirements.txt
# KAI Agent Requirements
# =======================
# Core dependencies for the KAI Agent system

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.10.0

# Vector Database and Knowledge Base
faiss-cpu>=1.7.4
pypdf>=3.17.0

# Docker Integration
docker>=7.0.0

# Configuration
python-dotenv>=1.0.0

# External APIs
tavily-python>=0.3.0
twilio>=8.10.0
elevenlabs>=0.2.0
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0

# Optional: PDF Generation (if needed)
reportlab>=4.0.0

# Note: Streamlit should already be installed in the main application
# If not, add: streamlit>=1.30.0